add a search documentation command and mcp tool
